Separate BabyMeal and LLM
